fix: guard CONTEXTUAL_RISK_FLAGS lookup against prototype pollution
Bracket-notation lookup on a plain object returned inherited prototype
methods (e.g. constructor, toString) which don't have .has(), causing a
TypeError at runtime. Guard with Object.hasOwn() so only own properties
are returned.

Extract normalization logic into scripts/normalize.mjs for testability
and add regression tests covering prototype property names.
